Fix bugs and add round tracking
