+++ src/Core/TabletopPowerTools.Core.csproj
@@ -13,6 +13,7 @@
 	  <PackageReference Include="MediatR" Version="11.1.0" />
 	  <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
 	  <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="7.0.1" />
+	  <PackageReference Include="MudBlazor" Version="6.1.8" />
 	  <PackageReference Include="sqlite-net-pcl" Version="1.8.116" />
 	  <PackageReference Include="SQLitePCLRaw.core" Version="2.1.3" />
 	  <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.0" />
