+++ src/Maui/TabletopPowerTools.Maui.csproj
@@ -51,7 +51,6 @@
         <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
         <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="7.0.0" />
         <PackageReference Include="Microsoft.Graph.Core" Version="2.0.14" />
-        <PackageReference Include="MudBlazor" Version="6.1.7" />
 
     </ItemGroup>
