+++ src/Maui/wwwroot/lib/game-icons/game-icons.css
@@ -12143,7 +12143,7 @@
 	content: "\fc9";
 }
 .game-icon-wingfoot:before {
-	content: "\fca";
+	content: "\ffca";
 }
 .game-icon-winter-gloves:before {
 	content: "\fcb";
